+++ examples/react-query/src/gen/hooks/pet/useUploadFileHook.ts
@@ -35,12 +35,13 @@ export async function uploadFileHook(
  * @summary uploads an image
  * {@link /pet/:petId/uploadImage}
  */
-export function useUploadFileHook(
+export function useUploadFileHook<TContext>(
   options: {
     mutation?: UseMutationOptions<
       UploadFileMutationResponse,
       ResponseErrorConfig<Error>,
-      { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams }
+      { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams },
+      TContext
     >
     client?: Partial<RequestConfig<UploadFileMutationRequest>> & { client?: typeof client }
   } = {},
@@ -51,7 +52,8 @@ export function useUploadFileHook(
   return useMutation<
     UploadFileMutationResponse,
     ResponseErrorConfig<Error>,
-    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams }
+    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams },
+    TContext
   >({
     mutationFn: async ({ petId, data, params }) => {
       return uploadFileHook({ petId }, data, params, config)