+++ examples/svelte-query/src/gen/hooks/createUploadFile.ts
@@ -35,12 +35,13 @@ export async function uploadFile(
  * @summary uploads an image
  * {@link /pet/:petId/uploadImage}
  */
-export function createUploadFile(
+export function createUploadFile<TContext>(
   options: {
     mutation?: CreateMutationOptions<
       UploadFileMutationResponse,
       ResponseErrorConfig<Error>,
-      { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams }
+      { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams },
+      TContext
     >
     client?: Partial<RequestConfig<UploadFileMutationRequest>> & { client?: typeof client }
   } = {},
@@ -51,7 +52,8 @@ export function createUploadFile(
   return createMutation<
     UploadFileMutationResponse,
     ResponseErrorConfig<Error>,
-    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams }
+    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams },
+    TContext
   >({
     mutationFn: async ({ petId, data, params }) => {
       return uploadFile(petId, data, params, config)