+++ examples/vue-query/src/gen/hooks/useCreateUsersWithListInput.ts
@@ -34,12 +34,13 @@ export async function createUsersWithListInput(
  * @summary Creates list of users with given input array
  * {@link /user/createWithList}
  */
-export function useCreateUsersWithListInput(
+export function useCreateUsersWithListInput<TContext>(
   options: {
     mutation?: MutationObserverOptions<
       CreateUsersWithListInputMutationResponse,
       ResponseErrorConfig<Error>,
-      { data?: MaybeRef<CreateUsersWithListInputMutationRequest> }
+      { data?: MaybeRef<CreateUsersWithListInputMutationRequest> },
+      TContext
     >
     client?: Partial<RequestConfig<CreateUsersWithListInputMutationRequest>> & { client?: typeof client }
   } = {},
@@ -47,7 +48,7 @@ export function useCreateUsersWithListInput(
   const { mutation: mutationOptions, client: config = {} } = options ?? {}
   const mutationKey = mutationOptions?.mutationKey ?? createUsersWithListInputMutationKey()
 
-  return useMutation<CreateUsersWithListInputMutationResponse, ResponseErrorConfig<Error>, { data?: CreateUsersWithListInputMutationRequest }>({
+  return useMutation<CreateUsersWithListInputMutationResponse, ResponseErrorConfig<Error>, { data?: CreateUsersWithListInputMutationRequest }, TContext>({
     mutationFn: async ({ data }) => {
       return createUsersWithListInput({ data }, config)
     },