+++ examples/vue-query/src/gen/hooks/useUpdateUser.ts
@@ -34,12 +34,13 @@ export async function updateUser(
  * @summary Update user
  * {@link /user/:username}
  */
-export function useUpdateUser(
+export function useUpdateUser<TContext>(
   options: {
     mutation?: MutationObserverOptions<
       UpdateUserMutationResponse,
       ResponseErrorConfig<Error>,
-      { username: MaybeRef<UpdateUserPathParams['username']>; data?: MaybeRef<UpdateUserMutationRequest> }
+      { username: MaybeRef<UpdateUserPathParams['username']>; data?: MaybeRef<UpdateUserMutationRequest> },
+      TContext
     >
     client?: Partial<RequestConfig<UpdateUserMutationRequest>> & { client?: typeof client }
   } = {},
@@ -47,7 +48,12 @@ export function useUpdateUser(
   const { mutation: mutationOptions, client: config = {} } = options ?? {}
   const mutationKey = mutationOptions?.mutationKey ?? updateUserMutationKey()
 
-  return useMutation<UpdateUserMutationResponse, ResponseErrorConfig<Error>, { username: UpdateUserPathParams['username']; data?: UpdateUserMutationRequest }>({
+  return useMutation<
+    UpdateUserMutationResponse,
+    ResponseErrorConfig<Error>,
+    { username: UpdateUserPathParams['username']; data?: UpdateUserMutationRequest },
+    TContext
+  >({
     mutationFn: async ({ username, data }) => {
       return updateUser({ username, data }, config)
     },