+++ examples/vue-query/src/gen/hooks/useUploadFile.ts
@@ -34,12 +34,13 @@ export async function uploadFile(
  * @summary uploads an image
  * {@link /pet/:petId/uploadImage}
  */
-export function useUploadFile(
+export function useUploadFile<TContext>(
   options: {
     mutation?: MutationObserverOptions<
       UploadFileMutationResponse,
       ResponseErrorConfig<Error>,
-      { petId: MaybeRef<UploadFilePathParams['petId']>; data?: MaybeRef<UploadFileMutationRequest>; params?: MaybeRef<UploadFileQueryParams> }
+      { petId: MaybeRef<UploadFilePathParams['petId']>; data?: MaybeRef<UploadFileMutationRequest>; params?: MaybeRef<UploadFileQueryParams> },
+      TContext
     >
     client?: Partial<RequestConfig<UploadFileMutationRequest>> & { client?: typeof client }
   } = {},
@@ -50,7 +51,8 @@ export function useUploadFile(
   return useMutation<
     UploadFileMutationResponse,
     ResponseErrorConfig<Error>,
-    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams }
+    { petId: UploadFilePathParams['petId']; data?: UploadFileMutationRequest; params?: UploadFileQueryParams },
+    TContext
   >({
     mutationFn: async ({ petId, data, params }) => {
       return uploadFile({ petId, data, params }, config)